css/css-animations/display-interpolation.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/6429">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<body>
<script>
const alwaysBlock = [
  {at: -1, expect: 'block'},
  {at: 0, expect: 'block'},
  {at: 0.1, expect: 'block'},
  {at: 0.9, expect: 'block'},
  {at: 1, expect: 'block'},
  {at: 1.5, expect: 'block'},
];
const alwaysNone = [
  {at: -1, expect: 'none'},
  {at: 0, expect: 'none'},
  {at: 0.1, expect: 'none'},
  {at: 0.9, expect: 'none'},
  {at: 1, expect: 'none'},
  {at: 1.5, expect: 'none'},
];
const blockToNone = [
  {at: -1, expect: 'block'},
  {at: 0, expect: 'block'},
  {at: 0.1, expect: 'block'},
  {at: 0.9, expect: 'block'},
  {at: 1, expect: 'none'},
  {at: 1.5, expect: 'none'},
];
const noneToBlock = [
  {at: -1, expect: 'none'},
  {at: 0, expect: 'none'},
  {at: 0.1, expect: 'block'},
  {at: 0.9, expect: 'block'},
  {at: 1, expect: 'block'},
  {at: 1.5, expect: 'block'},
];

test_no_interpolation({
  property: 'display',
  from: 'block',
  to: 'none',
  'CSS Transitions with transition-behavior:allow-discrete': blockToNone,
  'CSS Transitions with transition-property:all and transition-behavor:allow-discrete': blockToNone,
  'CSS Animations': blockToNone,
  'Web Animations': blockToNone
});

// This transitions test expect 'block' at every point because transitioning
// from display:none does not provide an initial style. This is expected and
// can be worked around by using @initial.
test_no_interpolation({
  property: 'display',
  from: 'none',
  to: 'block',
  'CSS Transitions with transition: all': alwaysBlock,
  'CSS Transitions': alwaysBlock,
  'CSS Transitions with transition-behavior:allow-discrete': alwaysBlock,
  'CSS Transitions with transition-property:all and transition-behavor:allow-discrete': alwaysBlock,
  'CSS Animations': noneToBlock,
  'Web Animations': noneToBlock
});

test_no_interpolation({
  property: 'display',
  from: 'inline',
  to: 'block'
});

test_no_interpolation({
  property: 'display',
  from: 'block',
  to: 'block'
});

test_no_interpolation({
  property: 'display',
  from: 'none',
  to: 'none'
}, alwaysNone);
</script>
